refactor(server): extract command preflight
