+++ server/proliferate/server/cloud/commands/service.py
@@ -19,17 +19,11 @@ from proliferate.constants.cloud import (
 )
 from proliferate.db.session_ops import is_integrity_error
 from proliferate.db.store import cloud_runtime_environments, cloud_workspaces
-from proliferate.db.store.cloud_agent_auth import store as agent_auth_store
-from proliferate.db.store.cloud_runtime_config import revisions as runtime_config_store
 from proliferate.db.store.cloud_sync import commands as commands_store
 from proliferate.db.store.cloud_sync import events as events_store
 from proliferate.db.store.cloud_sync import exposures as exposures_store
-from proliferate.db.store.cloud_sync import target_config as target_config_store
 from proliferate.db.store.cloud_sync import targets as targets_store
 from proliferate.server.cloud.claims.access import require_workspace_interact
-from proliferate.server.cloud.commands.agent_auth_refresh import (
-    queue_agent_auth_refresh_for_not_ready_preflight,
-)
 from proliferate.server.cloud.commands.domain.rules import (
     compact_command_json,
     validate_active_command_kind,
@@ -38,6 +32,13 @@ from proliferate.server.cloud.commands.domain.rules import (
     validate_command_source,
 )
 from proliferate.server.cloud.commands.models import CreateCloudCommandRequest
+from proliferate.server.cloud.commands.preflight import (
+    populate_agent_auth_preflight_payload,
+    stamp_managed_runtime_config_preflight,
+    validate_agent_auth_preflight,
+    validate_managed_runtime_config_current_for_command,
+    validate_runtime_config_preflight,
+)
 from proliferate.server.cloud.commands.wake import enqueue_managed_target_wake_outbox
 from proliferate.server.cloud.errors import CloudApiError
 from proliferate.server.cloud.event_logging import log_cloud_event
@@ -715,48 +716,6 @@ async def _resolve_cloud_workspace_id_for_target(
     return str(workspace.id)
 
 
-async def _populate_agent_auth_preflight_payload(
-    db: AsyncSession,
-    *,
-    target: targets_store.CloudTargetSnapshot,
-    kind: str,
-    payload: dict[str, object],
-) -> dict[str, object]:
-    if kind not in {
-        CloudCommandKind.start_session.value,
-        CloudCommandKind.send_prompt.value,
-    }:
-        return payload
-    if target.sandbox_profile_id is None:
-        next_payload = dict(payload)
-        next_payload.pop("agentAuthScope", None)
-        return next_payload
-    state = await agent_auth_store.get_target_state(
-        db,
-        sandbox_profile_id=target.sandbox_profile_id,
-        target_id=target.id,
-    )
-    if state is None:
-        profile = await agent_auth_store.get_sandbox_profile(db, target.sandbox_profile_id)
-        if profile is None:
-            return payload
-        required_revision = profile.agent_auth_revision
-    else:
-        required_revision = state.desired_revision
-    next_payload = dict(payload)
-    next_payload["sandboxProfileId"] = str(target.sandbox_profile_id)
-    next_payload["requiredAgentAuthRevision"] = required_revision
-    if kind == CloudCommandKind.start_session.value:
-        next_payload["agentAuthScope"] = {
-            "provider": "proliferate-cloud",
-            "id": str(target.sandbox_profile_id),
-            "targetId": str(target.id),
-        }
-    else:
-        next_payload.pop("agentAuthScope", None)
-    return next_payload
-
-
 def _direct_start_session_workspace_id(payload: dict[str, object]) -> str | None:
     value = payload.get("workspaceId")
     if not isinstance(value, str) or not value.strip():
@@ -843,20 +802,20 @@ async def enqueue_command(
         session_id=body.session_id,
         preconditions=body.preconditions,
     )
-    payload = await _populate_agent_auth_preflight_payload(
+    payload = await populate_agent_auth_preflight_payload(
         db,
         target=target,
         kind=kind,
         payload=body.payload,
     )
     validate_command_payload(kind=kind, payload=payload)
-    await _validate_agent_auth_preflight(
+    await validate_agent_auth_preflight(
         db,
         actor_user_id=user.id,
         target=target,
         payload=payload,
     )
-    payload = await _stamp_managed_runtime_config_preflight(
+    payload = await stamp_managed_runtime_config_preflight(
         db,
         actor_user_id=user.id,
         target=target,
@@ -869,13 +828,13 @@ async def enqueue_command(
         ),
     )
     validate_command_payload(kind=kind, payload=payload)
-    await _validate_runtime_config_preflight(
+    await validate_runtime_config_preflight(
         db,
         actor_user_id=user.id,
         target=target,
         payload=payload,
     )
-    await _validate_managed_runtime_config_current_for_command(
+    await validate_managed_runtime_config_current_for_command(
         db,
         actor_user_id=user.id,
         target=target,
@@ -1030,187 +989,6 @@ async def _record_pending_prompt_interaction_for_command(
     )
 
 
-async def _validate_agent_auth_preflight(
-    db: AsyncSession,
-    *,
-    actor_user_id: UUID,
-    target: targets_store.CloudTargetSnapshot,
-    payload: dict[str, object],
-) -> None:
-    sandbox_profile_id = payload.get("sandboxProfileId")
-    required_revision = payload.get("requiredAgentAuthRevision")
-    if sandbox_profile_id is None and required_revision is None:
-        return
-    try:
-        profile_id = UUID(str(sandbox_profile_id))
-    except (TypeError, ValueError) as exc:
-        raise CloudApiError(
-            "cloud_command_agent_auth_profile_invalid",
-            "Agent auth preflight sandboxProfileId is invalid.",
-            status_code=400,
-        ) from exc
-    if not isinstance(required_revision, int) or isinstance(required_revision, bool):
-        raise CloudApiError(
-            "cloud_command_agent_auth_revision_required",
-            "Agent auth preflight requiredAgentAuthRevision is required.",
-            status_code=400,
-        )
-    profile = await agent_auth_store.get_sandbox_profile(db, profile_id)
-    if profile is None:
-        raise CloudApiError(
-            "cloud_command_agent_auth_profile_not_found",
-            "Agent auth sandbox profile not found.",
-            status_code=404,
-        )
-    if profile.primary_target_id != target.id:
-        raise CloudApiError(
-            "cloud_command_agent_auth_target_mismatch",
-            "Agent auth sandbox profile is not attached to this target.",
-            status_code=409,
-        )
-    if profile.owner_scope == "personal":
-        if profile.owner_user_id != actor_user_id:
-            raise CloudApiError(
-                "cloud_command_agent_auth_profile_not_found",
-                "Agent auth sandbox profile not found.",
-                status_code=404,
-            )
-    elif profile.organization_id != target.organization_id or profile.organization_id is None:
-        raise CloudApiError(
-            "cloud_command_agent_auth_target_mismatch",
-            "Agent auth sandbox profile does not match this target organization.",
-            status_code=409,
-        )
-    if required_revision != profile.agent_auth_revision:
-        raise CloudApiError(
-            "cloud_command_agent_auth_revision_stale",
-            "Agent auth preflight revision is stale.",
-            status_code=409,
-        )
-    state = await agent_auth_store.get_target_state(
-        db,
-        sandbox_profile_id=profile.id,
-        target_id=target.id,
-    )
-    if (
-        state is None
-        or state.status != "applied"
-        or state.applied_revision is None
-        or state.applied_revision < required_revision
-    ):
-        await queue_agent_auth_refresh_for_not_ready_preflight(
-            sandbox_profile_id=profile.id,
-            target_id=target.id,
-            actor_user_id=actor_user_id,
-        )
-        raise CloudApiError(
-            "cloud_command_agent_auth_not_ready",
-            "Agent auth config has not been applied to this target.",
-            status_code=409,
-        )
-    if state.force_restart_required:
-        raise CloudApiError(
-            "cloud_command_agent_auth_restart_required",
-            "Agent auth changes require the session to restart before this command can run.",
-            status_code=409,
-        )
-
-
-async def _validate_runtime_config_preflight(
-    db: AsyncSession,
-    *,
-    actor_user_id: UUID,
-    target: targets_store.CloudTargetSnapshot,
-    payload: dict[str, object],
-) -> None:
-    sandbox_profile_id = payload.get("sandboxProfileId")
-    required_revision_id = payload.get("requiredRuntimeConfigRevisionId")
-    required_sequence = payload.get("requiredRuntimeConfigSequence")
-    required_content_hash = payload.get("requiredRuntimeConfigContentHash")
-    if (
-        sandbox_profile_id is None
-        and required_revision_id is None
-        and required_sequence is None
-        and required_content_hash is None
-    ):
-        return
-    try:
-        profile_id = UUID(str(sandbox_profile_id))
-    except (TypeError, ValueError) as exc:
-        raise CloudApiError(
-            "cloud_command_runtime_config_profile_invalid",
-            "Runtime config preflight sandboxProfileId is invalid.",
-            status_code=400,
-        ) from exc
-    if target.sandbox_profile_id != profile_id:
-        raise CloudApiError(
-            "cloud_command_runtime_config_target_mismatch",
-            "Runtime config sandbox profile is not attached to this target.",
-            status_code=409,
-        )
-    if target.owner_scope == "personal" and target.owner_user_id != actor_user_id:
-        raise CloudApiError(
-            "cloud_command_runtime_config_profile_not_found",
-            "Runtime config sandbox profile not found.",
-            status_code=404,
-        )
-    if not isinstance(required_revision_id, str) or not required_revision_id.strip():
-        raise CloudApiError(
-            "cloud_command_runtime_config_revision_required",
-            "Runtime config preflight revision id is required.",
-            status_code=400,
-        )
-    if not isinstance(required_sequence, int) or isinstance(required_sequence, bool):
-        raise CloudApiError(
-            "cloud_command_runtime_config_sequence_required",
-            "Runtime config preflight sequence is required.",
-            status_code=400,
-        )
-    if not isinstance(required_content_hash, str) or not required_content_hash.strip():
-        raise CloudApiError(
-            "cloud_command_runtime_config_hash_required",
-            "Runtime config preflight content hash is required.",
-            status_code=400,
-        )
-    _current, current_revision = await runtime_config_store.get_current(
-        db,
-        sandbox_profile_id=profile_id,
-    )
-    if current_revision is None:
-        raise CloudApiError(
-            "cloud_command_runtime_config_missing",
-            "Runtime config has not been compiled for this sandbox profile.",
-            status_code=409,
-        )
-    if (
-        str(current_revision.id) != required_revision_id
-        or current_revision.sequence != required_sequence
-        or current_revision.content_hash != required_content_hash
-    ):
-        raise CloudApiError(
-            "cloud_command_runtime_config_revision_stale",
-            "Runtime config preflight revision is stale.",
-            status_code=409,
-        )
-    _raise_runtime_config_blocked_if_needed(current_revision)
-    state = await agent_auth_store.get_target_state(
-        db,
-        sandbox_profile_id=profile_id,
-        target_id=target.id,
-    )
-    if (
-        state is None
-        or state.runtime_config_status != "applied"
-        or state.applied_runtime_config_revision_id != required_revision_id
-        or state.applied_runtime_config_sequence < required_sequence
-    ):
-        raise CloudApiError(
-            "cloud_command_runtime_config_not_ready",
-            "Runtime config has not been applied to this target.",
-            status_code=409,
-        )
-
-
 async def get_command_status(
     db: AsyncSession,
     *,
@@ -1352,260 +1130,6 @@ async def mark_pending_prompt_interaction_failed_for_command(
     command: commands_store.CloudCommandSnapshot,
 ) -> None:
     await _mark_pending_prompt_interaction_failed_for_command(db, command)
-
-
-async def _validate_managed_runtime_config_current_for_command(
-    db: AsyncSession,
-    *,
-    actor_user_id: UUID,
-    target: targets_store.CloudTargetSnapshot,
-    kind: str,
-    require_target_config: bool = True,
-) -> None:
-    if kind not in {
-        CloudCommandKind.decide_plan.value,
-        CloudCommandKind.resolve_interaction.value,
-        CloudCommandKind.update_session_config.value,
-        CloudCommandKind.cancel_turn.value,
-        CloudCommandKind.close_session.value,
-    }:
-        return
-    if target.sandbox_profile_id is None:
-        return
-    if target.owner_scope == "personal" and target.owner_user_id != actor_user_id:
-        raise CloudApiError(
-            "cloud_command_runtime_config_profile_not_found",
-            "Runtime config sandbox profile not found.",
-            status_code=404,
-        )
-    if require_target_config:
-        configs = await target_config_store.list_target_configs(db, target_id=target.id)
-        if not configs:
-            raise CloudApiError(
-                "cloud_command_target_config_required",
-                "Managed targets require a materialized target config before sessions can start.",
-                status_code=409,
-            )
-    _current, current_revision = await runtime_config_store.get_current(
-        db,
-        sandbox_profile_id=target.sandbox_profile_id,
-    )
-    if current_revision is None:
-        raise CloudApiError(
-            "cloud_command_runtime_config_missing",
-            "Runtime config has not been compiled for this sandbox profile.",
-            status_code=409,
-        )
-    _raise_runtime_config_blocked_if_needed(current_revision)
-    state = await agent_auth_store.get_target_state(
-        db,
-        sandbox_profile_id=target.sandbox_profile_id,
-        target_id=target.id,
-    )
-    if (
-        state is None
-        or state.runtime_config_status != "applied"
-        or state.applied_runtime_config_revision_id != str(current_revision.id)
-        or state.applied_runtime_config_sequence < current_revision.sequence
-    ):
-        raise CloudApiError(
-            "cloud_command_runtime_config_not_ready",
-            "Runtime config has not been applied to this target.",
-            status_code=409,
-        )
-
-
-async def _stamp_managed_runtime_config_preflight(
-    db: AsyncSession,
-    *,
-    actor_user_id: UUID,
-    target: targets_store.CloudTargetSnapshot,
-    kind: str,
-    payload: dict[str, object],
-    require_target_config: bool = True,
-) -> dict[str, object]:
-    del actor_user_id
-    if kind not in {
-        CloudCommandKind.start_session.value,
-        CloudCommandKind.send_prompt.value,
-        CloudCommandKind.decide_plan.value,
-    }:
-        return payload
-    if target.sandbox_profile_id is None:
-        return payload
-    if _runtime_config_preflight_fields_present(payload):
-        return payload
-    if require_target_config:
-        configs = await target_config_store.list_target_configs(db, target_id=target.id)
-        if not configs:
-            raise CloudApiError(
-                "cloud_command_target_config_required",
-                "Managed targets require a materialized target config before sessions can start.",
-                status_code=409,
-            )
-    _current, current_revision = await runtime_config_store.get_current(
-        db,
-        sandbox_profile_id=target.sandbox_profile_id,
-    )
-    if current_revision is None:
-        raise CloudApiError(
-            "cloud_command_runtime_config_missing",
-            "Runtime config has not been compiled for this sandbox profile.",
-            status_code=409,
-        )
-    _raise_runtime_config_blocked_if_needed(current_revision)
-    state = await agent_auth_store.get_target_state(
-        db,
-        sandbox_profile_id=target.sandbox_profile_id,
-        target_id=target.id,
-    )
-    if (
-        state is None
-        or state.runtime_config_status != "applied"
-        or state.applied_runtime_config_revision_id != str(current_revision.id)
-        or state.applied_runtime_config_sequence < current_revision.sequence
-    ):
-        raise CloudApiError(
-            "cloud_command_runtime_config_not_ready",
-            "Runtime config has not been applied to this target.",
-            status_code=409,
-        )
-    stamped = dict(payload)
-    stamped["sandboxProfileId"] = str(target.sandbox_profile_id)
-    stamped["requiredRuntimeConfigRevisionId"] = str(current_revision.id)
-    stamped["requiredRuntimeConfigSequence"] = current_revision.sequence
-    stamped["requiredRuntimeConfigContentHash"] = current_revision.content_hash
-    if kind == CloudCommandKind.start_session.value:
-        stamped["expectedRuntimeConfigRevision"] = {
-            "revisionId": str(current_revision.id),
-            "sequence": current_revision.sequence,
-            "contentHash": current_revision.content_hash,
-            "externalScope": {
-                "provider": "proliferate-cloud",
-                "id": str(target.sandbox_profile_id),
-                "targetId": str(target.id),
-            },
-        }
-    return stamped
-
-
-def _runtime_config_preflight_fields_present(payload: dict[str, object]) -> bool:
-    return any(
-        key in payload
-        for key in (
-            "requiredRuntimeConfigRevisionId",
-            "requiredRuntimeConfigSequence",
-            "requiredRuntimeConfigContentHash",
-        )
-    )
-
-
-async def stamp_and_validate_runtime_config_preflight(
-    db: AsyncSession,
-    *,
-    actor_user_id: UUID,
-    target_id: UUID,
-    kind: str,
-    payload: dict[str, object],
-) -> dict[str, object]:
-    target = await targets_store.get_target_by_id(db, target_id)
-    if target is None:
-        raise CloudApiError(
-            "cloud_command_target_not_found",
-            "Target not found.",
-            status_code=404,
-        )
-    stamped = await _stamp_managed_runtime_config_preflight(
-        db,
-        actor_user_id=actor_user_id,
-        target=target,
-        kind=kind,
-        payload=payload,
-    )
-    await _validate_runtime_config_preflight(
-        db,
-        actor_user_id=actor_user_id,
-        target=target,
-        payload=stamped,
-    )
-    return stamped
-
-
-async def stamp_and_validate_command_preflight(
-    db: AsyncSession,
-    *,
-    actor_user_id: UUID,
-    target_id: UUID,
-    kind: str,
-    payload: dict[str, object],
-) -> dict[str, object]:
-    target = await targets_store.get_target_by_id(db, target_id)
-    if target is None:
-        raise CloudApiError(
-            "cloud_command_target_not_found",
-            "Target not found.",
-            status_code=404,
-        )
-    payload = await _populate_agent_auth_preflight_payload(
-        db,
-        target=target,
-        kind=kind,
-        payload=payload,
-    )
-    validate_command_payload(kind=kind, payload=payload)
-    await _validate_agent_auth_preflight(
-        db,
-        actor_user_id=actor_user_id,
-        target=target,
-        payload=payload,
-    )
-    payload = await _stamp_managed_runtime_config_preflight(
-        db,
-        actor_user_id=actor_user_id,
-        target=target,
-        kind=kind,
-        payload=payload,
-    )
-    validate_command_payload(kind=kind, payload=payload)
-    await _validate_runtime_config_preflight(
-        db,
-        actor_user_id=actor_user_id,
-        target=target,
-        payload=payload,
-    )
-    await _validate_managed_runtime_config_current_for_command(
-        db,
-        actor_user_id=actor_user_id,
-        target=target,
-        kind=kind,
-    )
-    return payload
-
-
-def _raise_runtime_config_blocked_if_needed(
-    revision: runtime_config_store.SandboxProfileRuntimeConfigRevisionSnapshot,
-) -> None:
-    blocking_errors = _runtime_config_blocking_errors(revision.manifest_json)
-    if not blocking_errors:
-        return
-    raise CloudApiError(
-        "cloud_command_runtime_config_blocked",
-        "Runtime config has blocking resolver errors and cannot launch.",
-        status_code=409,
-    )
-
-
-def _runtime_config_blocking_errors(manifest_json: str) -> list[dict[str, object]]:
-    try:
-        manifest = json.loads(manifest_json)
-    except ValueError:
-        return []
-    if not isinstance(manifest, dict):
-        return []
-    blocking_errors = manifest.get("blockingErrors")
-    if not isinstance(blocking_errors, list):
-        return []
-    return [item for item in blocking_errors if isinstance(item, dict)]
 
 
 def _command_requires_managed_target_wake(
